requirements.txt
selenium==4.5.0
webdriver-manager==3.8.4
beautifulsoup4==4.11.1
requests==2.28.1
packaging
google-cloud-storage==2.7.0
instascrape
google-cloud-bigquery==3.5.0
google-cloud-vision==3.4.0
pandas
pandas-gbq
